[ Mobile ] Fixed mobile app not reporting webm MIME type
This change fixes a small BUG with the mobile app's `webm` handling
logic. Specifically, the app seems to correctly brows, play back, and
attempt to upload `webm` files to the server. However, it seems to not
correctly set the MIME type on the upload request resulting in the
server rejecting it.

The fix here seems to be pretty straight forward as we just need to add
a case statement to the file name -> mime type handling logic. As per
Mozilla's guide on MIME types, we always set `.webm`s to be
`video/webm`.