feat(web, server): Allow use of external geocoder service
Add the option to use an external (reverse) geocoder service.
The internal implementation is limited and sometimes returns inaccurate
results.
Instead of extending the current implementation, and increasing the
minimal hardware requirements, users should be able to use any service.
This allows users to self-host, i.e., Nominatim, use the Google Cloud
(#2684) or any other custom implementation.

The current drawback of the implementation is, that after the
configuration has been changed, the server must be restarted.