feat: Separate duplicate asset notification from upload success
This commit separates the notification messages for duplicate assets and upload success in the upload panel component. Previously, if a single duplicated image was uploaded, it erroneously displayed both the "Upload success" and "Duplicate" messages, leading to the invalidation of the successful message. Now, the "Upload success" message is displayed only when there are successful uploads other than duplicates. Subsequently, if there are any skipped duplicate assets, their count will be displayed in a separate warning message. This organization ensures that the notifications are displayed as desired.