Update backup-and-restore.md - added "-d immich" to psql restore command
The current command to unzip the backup and restore it to the database using the psql command malfunctions, because the database name "immich" needs to be included.

I've made several backups/restores now and none work if the "-d immich" is absent - all work just fine when it is present.

So, we need to update the instructions now.

I backed-up/restored using v1.104.0 - it's possible something works different in v1.105.x.  But regardless of that, many people restoring will be using versions earlier than 1.105.