Prevent losing photos
The text "Verplaatsen naar prullenbak" [move to bin] is too long. Only the first 2 words are visible suggesting that you 'move' photos elsewhere. If you do this with a selection in an album, the photos are removed from the album and good luck finding them again...
We fix this by replacing the text with "Naar prullenbak" which hopefully does fit.