Use STV-AV1 encoder for AV1.
STV-AV1 is much, much faster than the reference libaom encoder. In most real-world cases libaom is too slow to be practical, often struggling to achieve 0.1fps for 1080p content.

I suspect this is what was intended in f1ca1794a182d71e45340a262a450e694f4f1600 when AV1 support was originally added as it sets the `-svtav1-params` flag which is ignored unless using the STV-AV1 encoder. Furthermore there is lots of discussion about STV-AV1 in the PR thread.