Fix marking items as "not a duplicate".
The problem arose because `null == undefined` in JS. The other caparisons in this block should probably use `!==` as well (and `===` should be the general default because it is more predictable) but they are technically fine because no valid values will `== undefined`. So I have left them for now.

I am not sure if this is the best fix, but it does appear to solve the problem.

All credit goes to https://github.com/zzzorgo. I just converted it into a patch.

Fixes: https://github.com/immich-app/immich/issues/16413