fix(web): Handle upload failures from public users
When an album is shared via URL and public users are allowed to upload,
any errors they encounter are not correctly propogated to the UI. This
is due to the assumption that any uploads will be performed by
authenticated users and such if an error occurs, and there is no user
logged in it should be suppressed as they have logged out.

For the public user case we don't want to suppress these errors, so we
add a simple check to see if there was a user logged in *prior* to the
upload being initiated, and only suppress if the user is subsequently
not found.

In addition, we add tests to catch these cases to prevent regressions in
future.