docs: Document config options for hardware transcoding
The default configuration file values are already documented, but when
setting up hardware transcoding while using a config file, it's not easy
to figure out what options map to the settings, nor is it obvious what
the right values would be.

I've added a block that documents the equivalent options for the config
file based on what's described above. I picked `qsv` as my example and
added `nvenc` as another option.